requirements.txt
streamlit==1.32.0
openai==1.12.0
python-dotenv==1.0.1
nest-asyncio>=1.5.6
mcp>=0.1.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
langchain-core>=0.1.10
langchain-text-splitters>=0.0.1
faiss-cpu>=1.7.4
pymongo>=4.6.0
instructor>=0.4.0
pydantic>=2.0.0
watchdog>=3.0.0 